studymateApp/requirements/common.txt
django
djangorestframework
Pillow
pyjwt
django-cors-headers